Swap BigDecimal to float in "capital" field

+++ core/src/com/gpteam/shopmanager/company/Company.java
@@ -1,9 +1,7 @@
 package com.gpteam.shopmanager.company;
 
-import java.math.BigDecimal;
-
 public class Company {
     private String shortName;
     private String fullName;
-    private BigDecimal capital;
+    private float capital;
 }
